FarmerRecord factored out
